[Integrate] Bump LLVM to c22828991e7ca7b99048761c078252e94403ba6e (#34294)
Bump llvm. Required:
- updating a few APIs in the lldb formatters to use uint32_t
- adding a type convert for mo tensor -> mo tensor (thanks Brendan)
- removing an unused import (flaky) in the python lsp tests that is no
longer installed (not sure why this failure was triggered only by the
llvm bump)

---------

MODULAR_ORIG_COMMIT_REV_ID: 2da26af8ed1eafd1633482d61cbcb7f06908bc98

+++ KGEN/lib/MojoLLDB/Language/Formatters/MojoDecoratorBasedTypeFormatter.cpp
@@ -33,13 +33,13 @@ public:
     return m_backend.GetChildAtIndex(0, /*can_create=*/true);
   }
 
-  size_t CalculateNumChildren() override {
+  uint32_t CalculateNumChildren() override {
     if (!MightHaveChildren())
       return 0;
     return GetSyntheticValue()->GetNumChildren();
   }
 
-  lldb::ValueObjectSP GetChildAtIndex(size_t idx) override {
+  lldb::ValueObjectSP GetChildAtIndex(uint32_t idx) override {
     return GetSyntheticValue()->GetChildAtIndex(idx);
   }
